==> ProjectWhatsApp/1-Data Preprocessing.ipynb <==
# Read and analyse WhatsApp chat transcripts, and find some insights.
    
    Most chatty person.
    Sentiment analysis for each person.
    Most mentioned.
    High Profanity.
    Topics 

import os
import re

f = '/Users/rammurthys/Downloads'
filename = 'f9_chat.txt'

# members part of this Group chat
members = ['dpk', 'pjk', 'dir', 'srk', 'trk', 'svk', 'krk', 'kk', 'abi','ram','nkk']

phone_book = {'919789987036': 'srk'
              ,'919500566509': 'ram'
              ,'919663655116': 'trk'
              ,'919604854613': 'kk'
              ,'919946628663': 'pjk'
              ,'919840648198': 'dir'
              ,'919677200872': 'svk'
              ,'17164309669': 'dpk'
              ,'919944213734': 'krk'
              ,'918056193719': 'nkk'
              ,'917338813130': 'abi'
             }

# RegEx to parse the text
regexed = re.compile(r'[\W\d\sAMP]+\s+([\w\s]+):(.*)')

# Lets do the clean up
previous = None

content = open(os.path.join(f, filename))

member_content_book = {}

for i in content:
    if 'F9' in i:
        continue
    if regexed.search(i):
        key, val = regexed.search(i).groups()
        member_content_book[key]=member_content_book.get(key, [''])+[val.strip()+' ']
        previous = key
    else:
        member_content_book[previous]=member_content_book.get(previous, [''])+[i.strip()+' ']

# Create pickle file for each member transcript 
# !mkdir transcripts
import pickle

data = {}
for key, val in member_content_book.items():
    data[key]=[''.join(val)]
member_content_book = data

for key, val in member_content_book.items():
    with open('transcripts/'+key+'.txt', 'wb') as fp:
        pickle.dump(val, fp)


# Create a DataFrame
import pandas as pd
pd.set_option('max_colwidth', 150)

df = pd.DataFrame.from_dict(member_content_book).transpose()
df.columns = ['transcript']
df = df.sort_index()

additional_stop_words = [' lam ', ' da ', ' tat ', ' thats ', ' na ', ' adhu ', ' ll ', ' dai ', ' ur ', ' im ', ' u ',
                         ' ithu ', ' ve ', ' la ', ' ah ', ' omitted ', ' ku ', ' oda ', ' dhan ', ' va ', ' erukku ', ' like '
                         ' nu ', ' don ',' nee ' ' unnaku ', ' unaku ', ' enaku ', ' dan ', ' appo ', ' intha ', ' indha '
                         ,' ok ',' okay ']

def cleaning_round1(trans):
    text = re.sub(r"[\d+\W+_]", ' ', trans.lower())
    for i in additional_stop_words:
        text = re.sub(i, ' ', text)
    return text

round1 = lambda x: cleaning_round1(x)

data_clean1 = pd.DataFrame(df.transcript.apply(round1))
data_clean1

# Organize the data

Corpus development.

Doc-term matrix.

full_names = ['Deepak', 'KarthikK', 'KKeyan', 'Rambi', 'Riyas', 'Selva', 'Krishna', 'Ram', 'RajS', 'RajT']
df

# pickle this for later use
df.to_pickle('corpus.pkl')

# Document Term Matrix

from sklearn.feature_extraction.text import CountVectorizer

cv = CountVectorizer(stop_words='english')
data_cv = cv.fit_transform(data_clean1.transcript)
data_dtm1 = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
data_dtm1.index = data_clean1.index
data_dtm1

# pickle Document term matrix for later use
data_dtm1.to_pickle('dtm1.pkl')

# pickle data clean obj, and countvectorizer object too
data_clean1.to_pickle('data_clean1.pkl')
pickle.dump(cv, open('cv.pkl', 'wb'))

d = {}
for key, val in member_content_book.items():
    d[key]= {}
    for i,j in phone_book.items():
        if i in val[0]:
            d[key][j]=d[key].get(j, 0)+1
